requirements.txt
Flask==0.12
pymongo==3.4.0
pybase62==0.1.3
requests==2.12.4